requirements.txt
python-telegram-bot
openai-whisper
vosk
langdetect
spacy
nltk
gtts
fastapi
uvicorn
python-dotenv 